Update to only allow user information edit by logged in user for their own account.
Add functions logged_in_user and correct_user to UserController, which
must be validated before user edit or update functions can be called.

Add ability to store location to session so that users are re-directed
to the page they attempted to access on log-in, instead of default.

Add tests for all new functions