Add (non-functional) follow, unfollow, following, followers link
Add stats partial to render following, followers count. Add SCSS in
to style the partial. Add stats partial to the profile, home pages.

Add button to user page to show when user is not self.  Label 'follow'
or 'unfollow', dependent on the current relationship status. Add
_follow and _unfollow partials to build/ delete relationship. NB these
methods still need to be added (to relationship controller?)

Add integration tests for all the above. Add controller tests for next
section (user should only be able to navigate to followers/ following for
another user when logged in)