Add follow and unfollow functionality in the UI
Add relationships_controller to create and destroy relationships.

Add support for AJAX async updates in the follow and unfollow partials,
and incorporate this into the relationships controller. Add
create.js.erb and destroy.js.erb to asyncronously update the follow_form
button to style follow/ unfollow and to update the count of the users
on the page.

Add integration tests for ajax and standard follow/ unfollow actions.
Add relationship controller tests to validate a user must be logged in
to follow/ unfollow.